+++ docs/changelog.rst
@@ -40,6 +40,7 @@ Version 3.0.0
   * "Delete Chat Message" :const:`~twitchAPI.twitch.Twitch.delete_chat_message()`
   * "Send Chat Announcement" :const:`~twitchAPI.twitch.Twitch.send_chat_announcement()`
   * "Get Soundtrack Current Track" :const:`~twitchAPI.twitch.Twitch.get_soundtrack_current_track()`
+  * "Get Soundtrack Playlist" :const:`~twitchAPI.twitch.Twitch.get_soundtrack_playlist()`
 * Removed the folllowing deprecated Endpoints:
 
   * "Get Banned Event"